backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
openai>=0.28.0
python-dotenv>=1.0.0
redis>=4.5.5